+++ core/src/main/kotlin/net/corda/core/flows/BroadcastTransactionFlow.kt
@@ -2,7 +2,6 @@ package net.corda.core.flows
 
 import co.paralleluniverse.fibers.Suspendable
 import net.corda.core.identity.Party
-import net.corda.core.serialization.CordaSerializable
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.utilities.NonEmptySet
 
@@ -18,16 +17,12 @@ import net.corda.core.utilities.NonEmptySet
 @InitiatingFlow
 class BroadcastTransactionFlow(val notarisedTransaction: SignedTransaction,
                                val participants: NonEmptySet<Party>) : FlowLogic<Unit>() {
-    @CordaSerializable
-    data class NotifyTxRequest(val tx: SignedTransaction)
-
     @Suspendable
     override fun call() {
         // TODO: Messaging layer should handle this broadcast for us
-        val msg = NotifyTxRequest(notarisedTransaction)
         participants.filter { it != serviceHub.myInfo.legalIdentity }.forEach { participant ->
-            // This pops out the other side in NotifyTransactionHandler
-            send(participant, msg)
+            // SendTransactionFlow allows otherParty to access our data to resolve the transaction.
+            subFlow(SendTransactionFlow(participant, notarisedTransaction))
         }
     }
 }