+++ core/src/main/kotlin/net/corda/core/flows/ResolveTransactionsFlow.kt
@@ -1,9 +1,11 @@
 package net.corda.core.flows
 
 import co.paralleluniverse.fibers.Suspendable
-import net.corda.core.utilities.exactAdd
 import net.corda.core.crypto.SecureHash
 import net.corda.core.identity.Party
+import net.corda.core.internal.FetchAttachmentsFlow
+import net.corda.core.internal.FetchDataFlow
+import net.corda.core.internal.FetchTransactionsFlow
 import net.corda.core.serialization.CordaSerializable
 import net.corda.core.transactions.SignedTransaction
 import java.util.*
@@ -59,6 +61,7 @@ class ResolveTransactionsFlow(private val txHashes: Set<SecureHash>,
             require(result.size == transactions.size)
             return result
         }
+
     }
 
     @CordaSerializable
@@ -70,36 +73,37 @@ class ResolveTransactionsFlow(private val txHashes: Set<SecureHash>,
     // TODO: Figure out a more appropriate DOS limit here, 5000 is simply a very bad guess.
     /** The maximum number of transactions this flow will try to download before bailing out. */
     var transactionCountLimit = 5000
+        set(value) {
+            require(value > 0) { "$value is not a valid count limit" }
+            field = value
+        }
 
     @Suspendable
     @Throws(FetchDataFlow.HashNotFound::class)
     override fun call(): List<SignedTransaction> {
-        val newTxns: Iterable<SignedTransaction> = topologicalSort(downloadDependencies(txHashes))
+        // Start fetching data.
+        val newTxns = downloadDependencies(txHashes)
+        fetchMissingAttachments(newTxns + signedTransaction)
+        send(otherSide, FetchDataFlow.Request.End)
+        // Finish fetching data.
 
-        // For each transaction, verify it and insert it into the database. As we are iterating over them in a
-        // depth-first order, we should not encounter any verification failures due to missing data. If we fail
-        // half way through, it's no big deal, although it might result in us attempting to re-download data
-        // redundantly next time we attempt verification.
-        val result = ArrayList<SignedTransaction>()
-
-        for (stx in newTxns) {
-            // TODO: We could recover some parallelism from the dependency graph.
-            stx.verify(serviceHub)
-            serviceHub.recordTransactions(stx)
-            result += stx
+        val result = topologicalSort(newTxns).map {
+            // For each transaction, verify it and insert it into the database. As we are iterating over them in a
+            // depth-first order, we should not encounter any verification failures due to missing data. If we fail
+            // half way through, it's no big deal, although it might result in us attempting to re-download data
+            // redundantly next time we attempt verification.
+            it.verify(serviceHub)
+            serviceHub.recordTransactions(it)
+            it
         }
 
-        // If this flow is resolving a specific transaction, make sure we have its attachments as well
-        signedTransaction?.let {
-            fetchMissingAttachments(listOf(it))
-            result += it
-        }
-
-        return result
+        return signedTransaction?.let {
+            result + it
+        }?:result
     }
 
     @Suspendable
-    private fun downloadDependencies(depsToCheck: Set<SecureHash>): Collection<SignedTransaction> {
+    private fun downloadDependencies(depsToCheck: Set<SecureHash>): List<SignedTransaction> {
         // Maintain a work queue of all hashes to load/download, initialised with our starting set. Then do a breadth
         // first traversal across the dependency graph.
         //
@@ -121,35 +125,31 @@ class ResolveTransactionsFlow(private val txHashes: Set<SecureHash>,
         val resultQ = LinkedHashMap<SecureHash, SignedTransaction>()
 
         val limit = transactionCountLimit
-        check(limit > 0) { "$limit is not a valid count limit" }
         var limitCounter = 0
         while (nextRequests.isNotEmpty()) {
             // Don't re-download the same tx when we haven't verified it yet but it's referenced multiple times in the
             // graph we're traversing.
             val notAlreadyFetched = nextRequests.filterNot { it in resultQ }.toSet()
-            nextRequests.clear()
-
-            if (notAlreadyFetched.isEmpty())   // Done early.
-                break
+            if (notAlreadyFetched.isEmpty())
+                break // Done early.
 
             // Request the standalone transaction data (which may refer to things we don't yet have).
-            val downloads: List<SignedTransaction> = subFlow(FetchTransactionsFlow(notAlreadyFetched, otherSide)).downloaded
-
-            fetchMissingAttachments(downloads)
+            val downloads = subFlow(FetchTransactionsFlow(notAlreadyFetched, otherSide)).downloaded
 
             for (stx in downloads)
                 check(resultQ.putIfAbsent(stx.id, stx) == null)   // Assert checks the filter at the start.
 
             // Add all input states to the work queue.
-            val inputHashes = downloads.flatMap { it.inputs }.map { it.txhash }
+            val inputHashes = downloads.flatMap { it.tx.inputs }.map { it.txhash }
+
+            nextRequests.clear()
             nextRequests.addAll(inputHashes)
 
             limitCounter = limitCounter exactAdd nextRequests.size
             if (limitCounter > limit)
                 throw ExcessivelyLargeTransactionGraph()
         }
-
-        return resultQ.values
+        return resultQ.values.toList()
     }
 
     /**