+++ docs/source/example-code/src/main/kotlin/net/corda/docs/FlowCookbook.kt
@@ -63,6 +63,7 @@ object FlowCookbook {
                 // subflow's progress steps in our flow's progress tracker.
                 override fun childProgressTracker() = CollectSignaturesFlow.tracker()
             }
+
             object VERIFYING_SIGS : Step("Verifying a transaction's signatures.")
             object FINALISATION : Step("Finalising a transaction.") {
                 override fun childProgressTracker() = FinalityFlow.tracker()
@@ -384,12 +385,12 @@ object FlowCookbook {
             // dependencies, we'd need to download all of these dependencies
             // using``ResolveTransactionsFlow`` before verifying it.
             // DOCSTART 13
-            subFlow(ResolveTransactionsFlow(twiceSignedTx, counterparty))
+            subFlow(ResolveTransactionsFlow(counterparty, twiceSignedTx))
             // DOCEND 13
 
             // We can also resolve a `StateRef` dependency chain.
             // DOCSTART 14
-            subFlow(ResolveTransactionsFlow(setOf(ourStateRef.txhash), counterparty))
+            subFlow(ResolveTransactionsFlow(counterparty, setOf(ourStateRef.txhash)))
             // DOCEND 14
 
             // A ``SignedTransaction`` is a pairing of a ``WireTransaction``